Payment for Registration

+++ backend/eventhub-backend/src/main/java/com/eventhub/payment/service/PaymentService.java
@@ -5,5 +5,5 @@ import com.eventhub.payment.dto.response.PaymentResponse;
 
 public interface PaymentService {
 
-	PaymentResponse payForRegistration(Long registrationId,PaymentRequest request);
+	PaymentResponse payForRegistration(Integer registrationId,PaymentRequest request);
 }
